serious refactoring going on here

+++ lib/dayone.rb
@@ -1,53 +1,14 @@
-class DayOne < SocialLogger
-  def initialize(options = {})
-    options['storage'] ||= 'icloud'
-    if options['storage'].downcase == 'icloud'
-      dayonedir = %x{ls ~/Library/Mobile\\ Documents/|grep dayoneapp}.strip
-      full_path = File.expand_path("~/Library/Mobile\ Documents/#{dayonedir}/Documents/Journal_dayone/")
-      if File.exists?(full_path)
-        @dayonepath = full_path
-      else
-        raise "Failed to find iCloud storage path"
-        Process.exit(-1)
-      end
-    elsif File.exists?(File.expand_path(options['storage']))
-      @dayonepath = File.expand_path(options['storage'])
-    else
-      raise "Path not specified or doesn't exist: #{options['storage']}"
-      Process.exit(-1)
-    end
-
-    @template = ERB.new <<-XMLTEMPLATE
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-  <key>Creation Date</key>
-  <date><%= datestamp %></date>
-  <key>Entry Text</key>
-  <string><%= entry %></string>
-  <key>Starred</key>
-  <<%= starred %>/>
-  <key>UUID</key>
-  <string><%= uuid %></string>
-</dict>
-</plist>
-XMLTEMPLATE
-  end
-  attr_accessor :dayonepath
-
+class DayOne < Slogger
   def to_dayone(options = {})
     content = options['content'] || ''
     uuid = options['uuid'] || %x{uuidgen}.gsub(/-/,'').strip
     starred = options['starred'] || false
     datestamp = options['datestamp'] || Time.now.utc.iso8601
 
-    if @debug || options['debug']
-      return true
-    end
-
     # entry = CGI.escapeHTML(content.unpack('C*').pack('U*').gsub(/[^[:punct:]\w\s]+/,' ')) unless content.nil?
     entry = CGI.escapeHTML(content) unless content.nil?
+    @dayonepath = storage_path
+    @log.info("Saving entry to #{@dayonepath}/entries/#{uuid}.doentry")
     fh = File.new(File.expand_path(@dayonepath+'/entries/'+uuid+".doentry"),'w+')
     fh.puts @template.result(binding)
     fh.close
@@ -55,18 +16,20 @@ XMLTEMPLATE
   end
 
   def save_image(imageurl,uuid)
+    @dayonepath = Slogger.new.storage_path
     source = imageurl.gsub(/^https/,'http')
     match = source.match(/(\..{3,4})($|\?|%22)/)
     unless match.nil?
       ext = match[1]
     else
-      Logger.new(STDERR).warn("Attempted to save #{imageurl} but extension could not be determined")
+      @log.warn("Attempted to save #{imageurl} but extension could not be determined")
       ext = '.jpg'
     end
     target = @dayonepath + '/photos/'+uuid+ext
     begin
       Net::HTTP.get_response(URI.parse(imageurl)) do |http|
         data = http.body
+        @log.info("Retrieving image: #{imageurl} to #{target}")
         open( File.expand_path(target), "wb" ) { |file| file.write(data) }
       end
       return self.process_image(target)
@@ -81,6 +44,7 @@ XMLTEMPLATE
     match = orig.match(/(\..{3,4})$/)
     return false if match.nil?
     ext = match[1]
+    @log.info("Resizing image #{orig}")
     %x{sips -Z 800 "#{orig}"}
     unless ext =~ /\.jpg$/
       case ext
@@ -89,10 +53,16 @@ XMLTEMPLATE
         FileUtils.mv(orig,target)
         return target
       when /\.(png|gif)$/
-        target = orig.gsub(/#{ext}$/,'.jpg')
-        %x{/usr/local/bin/convert "#{orig}" "#{target}"}
-        File.delete(orig)
-        return target
+        if File.exists?('/usr/local/bin/convert')
+          target = orig.gsub(/#{ext}$/,'.jpg')
+          @log.info("Converting #{orig} to JPEG")
+          %x{/usr/local/bin/convert "#{orig}" "#{target}"}
+          File.delete(orig)
+          return target
+        else
+          @log.warn("Image could not be converted to JPEG format and may not show up in Day One. Please install ImageMagick.")
+          return orig
+        end
       else
         return orig
       end
@@ -106,7 +76,7 @@ XMLTEMPLATE
     options['uuid'] ||= %x{uuidgen}.gsub(/-/,'').strip
     options['starred'] ||= false
     options['datestamp'] ||= Time.now.utc.iso8601
-    target_path = File.expand_path(@dayonepath+"/photos/"+options['uuid']+".jpg")
+    target_path = File.expand_path(Slogger.new.storage_path+"/photos/"+options['uuid']+".jpg")
 
     if copy
       FileUtils.copy(File.expand_path(file),target_path)
